+++ test/integration/failover-compose.test.ts
@@ -12,7 +12,7 @@ import * as os from "node:os";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import { readFileSync } from "node:fs";
-import ModelRouterPlugin from "../../src/index";
+import { ModelRouterPluginV1 as ModelRouterPlugin } from "../../src/index";
 import { invalidateConfigCache, validateConfig } from "../../src/router/config";
 import { assembleSystemPrompt } from "../../src/router/protocol";